refactor: move minedtransaction to storage
